+++ lib/components/primitive-components/Group/Group_doInitialSchematicTraceRender/extendTraceEndpointsToReachPinsInsideExpandedBoundingBox.ts
